+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ### Added
 - New plugin: `json_ld` for generating JSON-LD tags in the output [#453]
 - New plugin: `purgecss` to remove unused CSS. [#693]
+- New middleware `router`.
 
 ## [2.4.3] - 2024-12-11
 ### Added